feat: add support for threads
This commit adds intial support for tweeting threads.

The idea is to separate tweets in a thread with ---

If the dashes are seen, the tweet logic will loop through the tweets one
after another.

1. Send the first tweet.
2. Wait for the response and extract the tweet id
3. Use that id to send the next tweet as a reply to that.

The current state is rough and just POC. The codebase is very dirty.

------
This feature is supported both in cli mode and server mode.